feat: #526 organize Manifest
